\name{BerkeleyUrls}
\alias{BerkeleyUrls}
\docType{data}
\title{ a data.frame of Berkeley URLS and zip filenames
 
}
\description{This data.frame contians the urls for the final Berkeley
 data. It also has zip file names. It is used to automate
 downloading via the function \code{downloadBerkeley}
 
}
\usage{ BerkeleyUrls }
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Url}}{a character vector}
    \item{\code{Zip}}{a character vector}
  }
}
\details{ The urls and zip names for the files.
  
 
}
\source{
  \url{http://berkeleyearth.org/data/}
}
 
 
\keyword{datasets}
