\name{isBigMatrix}
\alias{isBigMatrix}
 
\title{ isBigMatrix tests the class of an object
 
}
\description{Simply tests whether a matrix is a Big memory file backed 
 big matrix
 
}
\usage{
isBigMatrix(x)
}
 
\arguments{
  \item{x}{
   A matrix to be tested for its class
}
}
\details{ Fucntions such as removeNAStations test the type of data
 they are working with
 
}
\value{ Returns true false
 
}
 
\author{Steven Mosher
 
}
 

 
\keyword{Timeseries }
 
