\name{readSiteFlags}
\alias{readSiteFlags}
 
\title{A function to read site flags
 
}
\description{ The function reads the site flags assocaited with every station
 
}
\usage{
readSiteFlags(Directory, filename = "site_flags.txt")
}
 
\arguments{
  \item{Directory}{Source directory for the file
 
}
  \item{filename}{Default name is the name provided by berkeley
  "site_flags.txt"
 
}
}
\details{ The file contains flags for every station
 
}
\value{ returns a dataframe of the site Id and its associated
flags.
 
}
 
\author{Steven Mosher
 
}
 

 
 
\keyword{ stationdata }
 
